Up and running with Panel View.
No longer uses Panel (since only contained shape).
Add constants to the coordinates since they are messed up oops.
Add custom app to manifest!

+++ app/src/test/java/com/example/climbon/ExampleUnitTest.java
@@ -27,10 +27,10 @@ public class ExampleUnitTest {
 
     @Test
     public void genIrregularShape() {
-        ArrayList<Float> corners = new ArrayList<Float>(Arrays.<Float>asList((float) 0,(float) 0,(float) 10,(float) 0,(float) 10,(float) 8, (float) 8,(float) 10,(float) 0, (float) 10));
+        ArrayList<Float> corners = new ArrayList<Float>(Arrays.<Float>asList(0f,0f,10f,0f,10f,8f,8f,10f,0f,10f));
         Shape simple_shape;
         try{
-            simple_shape = new Shape(corners, new Coordinate((float) 0.5, (float) 0.5));
+            simple_shape = new Shape(corners, new Coordinate(0.5f,0.5f));
             assertEquals(15, simple_shape.getNumHolds());
         } catch (Exception e) {
             e.printStackTrace();
